Maman 11: Update Host Tests to 26.04
Updated the infrastructure, Dockerfiles and host tests to begin the
update to 26.04.

Tests:
As the build tools were updated to 26.04 versions, new warnings have
appeared, fixed all host test warnings.

Infrastructure:
A lot of infra scripts were changed to support a multi docker image
infrastructure: `compile-guest-tests.sh`, `compile-host-tests.sh`,
`docker-run-bash.sh`, `docker-run-sanity.sh`, `docker-test-guest.sh`,
and `docker-test-exofs.sh`  take a version argument (14.04/26.04); And
`build-qemu-image.sh`, `compile-kernel.sh`, `compile-qemu.sh`, and
`docker-test-host.sh` take version from `config.json`.
`build-docker-image.sh` builds all versions in `versions/containers/`.

Docker:
Added an 26.04 Dockerfile to `versions/containers/` adapter from 14.04
Dockerfile with fixes (mainly for package/lib versions). Docker images
are no longer built as `evssim:latest`, now every Dockerfile creates a
docker image with a different tag (same as the name of it's folder).

Changes:
versions/containers/26.04/Dockerfile [NEW]:
+ Added a new `Dockerfile` for an Ubuntu 26.04 based eVSSIM container.
* Compared to the 14.04 `Dockerfile`, update names package list to
  install. Some packages were renamed, some deprecated.
* Compared to the 14.04 `Dockerfile`, GTest stopped compiling because of
  minimum cmake version. Updated `cmake_minimum_required()` from 2.6.2
  to 3.5 in CMakeLists.txt.
* Compared to 14.04, libwebsockets v4.1 no longer compiles due to newer
  versions of OpenSSL. Updated to v4.5, and added
  `Wno-error=discarded-qualifiers` flag with `-DCMAKE_C_FLAGS` to the
  `cmake` command to avoid compile warnings(that fail the compilation).

infra/builder/build-docker-image.sh:
* Now builds a docker image for each container in the containers
  directory.

eVSSIM/CONFIG/vssim_config_manager.c:
* Fixed `stringop-overflow` warning in `GET_DATA_FILENAME()` function
  by using `snprintf` instead of `strncat`.

eVSSIM/FTL_SOURCE/PAGE_MAP/TOOLS/uthash.h:
* Pulled new version of uthash because Host Tests wouldn't compile with
  the old version.

eVSSIM/FTL_SOURCE/PAGE_MAP/ftl_gc_manager.h,
eVSSIM/LOG_MGR/logging_manager.c,
eVSSIM/LOG_MGR/logging_parser.c,
eVSSIM/osc-osd/osd-target/osd.c,
eVSSIM/osc-osd/osd-target/osd.h,
eVSSIM/tests/host/unit/test_onfi_ops.cc

eVSSIM/tests/host/build.mk:
* Changed `:=` to `+=` for the environment variables.
* Changed imports of external libraries from `-I` to `-isystem` to
  avoid warnings in them breaking compilation.

eVSSIM/tests/host/simulation/base_emulator_tests.h:
* Added `maybe_unused` attribute to `operator<<` override to avoid
  `unused-function` warning.
* Fixed `overloaded-virtual` warning in `TearDown` function overload
  from GTest. In GTest the function was `virtual void TearDown()`, but
  in `BaseTest` in this header, the function was overloaded as `virtual
  void TearDown(bool term_ssd_conf=true)`, and that created a conflict.
  Split the function into `BaseTearDown(bool term_ssd_conf=true)` and a
  new `TearDown()` override that simply calls `BaseTearDown(true)` when
  called.

eVSSIM/tests/host/simulation/gc_tests.cc,
eVSSIM/tests/host/simulation/log_mgr_tests.cc,
eVSSIM/tests/host/simulation/object_tests.cc,
eVSSIM/tests/host/simulation/offline_logger_tests.cc,
eVSSIM/tests/host/simulation/sector_tests.cc,
eVSSIM/tests/host/simulation/ssd_io_emulator_tests.cc,
eVSSIM/tests/host/simulation/ssd_program_compatible_test.cc,
eVSSIM/tests/host/simulation/ssd_write_read_test.cc:
* Fixed `overloaded-virtual` warning by changing the call to `TearDown`
  into the new `BaseTearDown()`. See change for `base_emulator_tests`.

eVSSIM/tests/host/simulation/onfi_ops_test.cc:
* Fixed `overloaded-virtual` warning by changing the call to `TearDown`
  into the new `BaseTearDown()`. See change for `base_emulator_tests`.
* Fixed 'maybe-uninitialized' warning in multiple tests due to the
  `buffer` array being declared but never initialized(Even though it is
  never used because the test is expected to error out, but GCC can't
  know that). To fix, added an empty initializer `={}`.

infra/builder/versions/containers/14.04/logging_server.c [NEW],
infra/builder/versions/containers/14.04/logging_server.h [NEW],
infra/builder/versions/containers/26.04/logging_server.c [NEW],
infra/builder/versions/containers/26.04/logging_server.h [NEW]:
* `log_server` depends on libwebsocket. On 14.04 libsockets version is
  v4.1, and on 26.04 it's version is v4.5. `log_server` is being used
  both by parts of the project that are on 14.04 and parts of the
  project that are 26.04, and due to changes between v4.1 and v4.5, two
  different versions are needed, one for each version.
* Both versions originate from `eVSSIM/MONITOR/SERVER/logging_server`.

infra/builder/build-qemu-image.sh:
* Gets version for the image from `configs.json`, but for now a
  hardcoded 14.04 is used.

infra/builder/compile-host-tests.sh:
* Gets version for the image from `configs.json`.
* Added a docker image version argument.
* Syntax for the `bear` command changed, now requiring `--`.
* Copy & Restore functionality for `log_server.c` and `log_server.h` as
  mentioned before, for compiling host and guest tests with the new
  `log_server` for libwebsocket v4.5 for 26.04.
* Added `-j` option to `make` for speeding up compilation.

infra/builder/compile-guest-tests.sh:
* Gets version for the image from `configs.json`.
* Added a docker image version argument.
* Copy & Restore functionality for `log_server.c` and `log_server.h` as
  mentioned before, for compiling host and guest tests with the new
  `log_server` for libwebsocket v4.5 for 26.04.
* Added `-j` option to `make` for speeding up compilation.

infra/builder/compile-kernel.sh:
* Gets docker image version from `configs.json`.

infra/builder/compile-qemu.sh:
* Gets docker image version from `configs.json`.
+ Added a `make clean` at the beginning because `osc-cd` is also
  compiled (but with different tooling) for the tests.

infra/builder/docker-run-bash.sh,
infra/builder/docker-run-sanity.sh,
infra/builder/docker-test-host.sh,
infra/builder/docker-test-guest.sh,
infra/builder/docker-test-exofs.sh:
* Gets docker image version from `configs.json`.

infra/builder/env.sh:
+ Added `EVSSIM_CONTAINERS_FOLDER` environmental variable - The location
  of the `versions/containers/` directory.

infra/builder/load_config.sh:
* Changed configuration to 5.

infra/builder/run-ci.sh:
* Added version parameters to some of the steps.

+++ eVSSIM/FTL_SOURCE/PAGE_MAP/ftl_gc_manager.h
@@ -9,7 +9,7 @@
 #include "ftl.h"
 
 typedef ftl_ret_val (*gc_collection_algo)(uint8_t, int, bool background);
-typedef ftl_ret_val (*gc_next_page_algo)(uint8_t, int, int, uint64_t*);
+typedef ftl_ret_val (*gc_next_page_algo)(uint8_t, int, uint64_t, uint64_t*); /* Corrected the third parameter to uint64_t */
 
 typedef struct gc_thread {
     pthread_t tid;
